requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pandas==2.1.3
rapidfuzz==3.5.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
jinja2==3.1.2
aiofiles==23.2.1